MAT-9655 move admin end points to AdminController
